chore: init public/protected selection button
