+++ frontend/src/views/courses/search/ProtectedSearchCourseView.vue
@@ -1,0 +1,125 @@
+<script setup lang="ts">
+import Accordion from 'primevue/accordion';
+import AccordionTab from 'primevue/accordiontab';
+import InputText from 'primevue/inputtext';
+import IconField from 'primevue/iconfield';
+import InputIcon from 'primevue/inputicon';
+import Checkbox from 'primevue/checkbox';
+import Paginator from 'primevue/paginator';
+import Title from '@/components/layout/Title.vue';
+import BaseLayout from '@/components/layout/base/BaseLayout.vue';
+import CourseGeneralList from '@/components/courses/CourseGeneralList.vue';
+import { onMounted } from 'vue';
+import { useCourses } from '@/composables/services/course.service.ts';
+import { useAuthStore } from '@/store/authentication.store.ts';
+import { useFaculty } from '@/composables/services/faculty.service.ts';
+import { storeToRefs } from 'pinia';
+import { useI18n } from 'vue-i18n';
+import { useFilter } from '@/composables/filters/filter.ts';
+import { usePaginator } from '@/composables/filters/paginator.ts';
+import { useRoute } from 'vue-router';
+import { getCourseFilters } from '@/types/filter/Filter.ts';
+import { getAcademicYears } from '@/types/Course.ts';
+
+/* Composable injections */
+const { t } = useI18n();
+const { query } = useRoute();
+const { user } = storeToRefs(useAuthStore());
+const { faculties, getFaculties } = useFaculty();
+const { pagination, searchCourses } = useCourses();
+const { onPaginate, resetPagination, page, first, pageSize } = usePaginator();
+const { filter, onFilter } = useFilter(getCourseFilters(query));
+
+/**
+ * Fetch the courses based on the filter.
+ */
+async function fetchCourses(): Promise<void> {
+    await searchCourses(filter.value, page.value, pageSize.value);
+}
+
+/* Fetch the faculties */
+onMounted(async () => {
+    // Fetch the faculties
+    await getFaculties();
+
+    /* Search courses on page change */
+    onPaginate(fetchCourses);
+
+    /* Search courses on filter change */
+    onFilter(fetchCourses);
+
+    /* Reset pagination on filter change */
+    onFilter(
+        async () => {
+            await resetPagination([pagination]);
+        },
+        0,
+        false,
+    );
+});
+</script>
+
+<template>
+    <BaseLayout>
+        <div class="grid">
+            <div class="col-12 xl:col-3">
+                <Accordion :active-index="[0]" multiple>
+                    <AccordionTab :header="t('views.courses.search.search')">
+                        <IconField iconPosition="left">
+                            <InputText
+                                :placeholder="t('views.courses.search.placeholder')"
+                                v-model="filter.search"
+                                class="w-full"
+                            />
+                            <InputIcon class="pi pi-search"></InputIcon>
+                        </IconField>
+                    </AccordionTab>
+                    <AccordionTab :header="t('views.courses.search.faculty')" v-if="faculties">
+                        <div v-for="faculty in faculties" :key="faculty.id" class="flex align-items-center mb-2">
+                            <Checkbox
+                                v-model="filter.faculties"
+                                :inputId="faculty.id"
+                                name="faculties"
+                                :value="faculty.id"
+                            />
+                            <label :for="faculty.id" class="ml-2 text-sm">{{ faculty.name }}</label>
+                        </div>
+                    </AccordionTab>
+                    <AccordionTab :header="t('views.courses.search.year')" v-if="user && pagination !== null">
+                        <div
+                            v-for="year in getAcademicYears(pagination.min_year, pagination.max_year)"
+                            :key="year"
+                            class="flex align-items-center mb-2"
+                        >
+                            <Checkbox
+                                v-model="filter.years"
+                                :inputId="year.toString()"
+                                name="faculties"
+                                :value="year"
+                            />
+                            <label :for="year.toString()" class="ml-2 text-sm"> {{ year }} - {{ year + 1 }} </label>
+                        </div>
+                    </AccordionTab>
+                </Accordion>
+            </div>
+            <div class="col-12 xl:col-9 mt-4 xl:mt-0">
+                <div class="flex justify-content-between align-items-center">
+                    <Title>
+                        {{ t('views.courses.search.title') }}
+                    </Title>
+                    <slot name="publicButton" />
+                </div>
+
+                <p class="mt-3" v-if="pagination">
+                    {{ t('views.courses.search.results', [pagination.count]) }}
+                </p>
+
+                <CourseGeneralList class="mt-3" :courses="pagination?.results ?? null" :cols="3" />
+
+                <Paginator :rows="pageSize" :total-records="pagination?.count" v-model:first="first" />
+            </div>
+        </div>
+    </BaseLayout>
+</template>
+
+<style scoped lang="scss"></style>
